post, like, followの操作に認証を追加

+++ backend/python/api/routers/posts.py
@@ -5,12 +5,13 @@ from schemas.user import UserRequest
 sys.path.append('../')
 
 from datetime import datetime
-from typing import List,Union
+from typing import Dict, List,Union
 from fastapi import APIRouter,Depends, HTTPException, status
 from schemas.post import Post, PostCreate, PostResponse
 
 from db import get_db
 from cruds.post import delete_post, get_all_posts, get_one_post, create_post, get_posts_by_user, get_posts_by_following_user
+from cruds.security import get_current_username_and_id
 
 router=APIRouter()
 
@@ -21,7 +22,8 @@ def list_posts(db=Depends(get_db)):
     return get_all_posts(db=db)
 
 @router.post("/posts", response_model=PostResponse)
-def users_new_post(post_body: PostCreate, db=Depends(get_db)):
+def users_new_post(post_body: PostCreate, db=Depends(get_db),userdata:Dict=Depends(get_current_username_and_id)):
+    post_body.posted_by=userdata.get("user_id")
     return create_post(db=db, post_create=post_body)
 
 @router.get("/posts/following", response_model=List[PostResponse])
@@ -37,5 +39,8 @@ def show_one_post_by_user(user_id:int,post_id:int, db=Depends(get_db)):
     return get_one_post(db=db, post_id=post_id)
 
 @router.delete("/posts/{user_id}/{post_id}", response_model=PostResponse)
-def delete_one_post(user_id:str,post_id:int, db=Depends(get_db)):
+def delete_one_post(user_id:int,post_id:int, db=Depends(get_db),userdata:Dict=Depends(get_current_username_and_id)):
+    if user_id!=userdata.get("user_id") and userdata.get("username")!="admin":
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED,detail="action is not authorized")
+
     return delete_post(db=db,user_id=user_id,post_id=post_id)
